requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
gradio
faiss-cpu
pymupdf
docx2txt
